Campaign service merge

+++ campaign-service/src/main/java/com/oem/evwarranty/repository/AppointmentRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
-    Page<Appointment> findByCampaignId(Integer campaignId, Pageable pageable);
-    Page<Appointment> findByAffectedId(Integer affectedId, Pageable pageable);
-    Page<Appointment> findByCampaignIdAndStatus(Integer campaignId, AppointmentStatus status, Pageable pageable);
-    Page<Appointment> findByAffected_Id(Integer affectedId, Pageable pageable);
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+    Page<Appointment> findByCampaignId(Long campaignId, Pageable pageable);
+    Page<Appointment> findByAffectedId(Long affectedId, Pageable pageable);
+    Page<Appointment> findByCampaignIdAndStatus(Long campaignId, AppointmentStatus status, Pageable pageable);
+    Page<Appointment> findByAffected_Id(Long affectedId, Pageable pageable);
 }
